chore: Fixed data type for Postion

+++ src/main/java/com/vocalink/crossproduct/domain/position/ParticipantPosition.java
@@ -1,7 +1,7 @@
 package com.vocalink.crossproduct.domain.position;
 
 import com.vocalink.crossproduct.ui.dto.position.ParticipantPositionDto;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,9 +15,9 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ParticipantPosition {
   private String participantId;
-  public BigInteger credit;
-  private BigInteger debit;
-  private BigInteger netPosition;
+  public BigDecimal credit;
+  private BigDecimal debit;
+  private BigDecimal netPosition;
 
   public ParticipantPositionDto toDto(){
     return ParticipantPositionDto.builder()
